rename enum API_KIND to api_kind to avoid conflict to a defined m4 macro
git-svn-id: https://svn.mcs.anl.gov/repos/parallel-netcdf/branches/am@2927 c368ceb8-fd28-0410-a457-e46ee1d7cd4a

+++ src/drivers/ncmpi/dim.c
@@ -460,7 +460,7 @@ err_check:
             return ncmpii_handle_error(mpireturn, "MPI_Bcast root_name_len");
         }
 
-        root_name = (char*) NCI_Malloc(root_name_len);
+        root_name = (char*) NCI_Malloc((size_t)root_name_len);
         root_name[0] = '\0';
         if (name != NULL) strcpy(root_name, name);
         TRACE_COMM(MPI_Bcast)(root_name, root_name_len, MPI_CHAR, 0, ncp->nciop->comm);
@@ -672,7 +672,7 @@ err_check:
             return ncmpii_handle_error(mpireturn, "MPI_Bcast root_name_len");
         }
 
-        root_name = (char*) NCI_Malloc(root_name_len);
+        root_name = (char*) NCI_Malloc((size_t)root_name_len);
         root_name[0] = '\0';
         if (newname != NULL) strcpy(root_name, newname);
         TRACE_COMM(MPI_Bcast)(root_name, root_name_len, MPI_CHAR, 0, ncp->nciop->comm);
